--- a/simd-fuzz-target/.gitignore
+++ b/simd-fuzz-target/.gitignore
@@ -1,3 +1,0 @@
-/target
-**/*.rs.bk
-out