--- a/simd-fuzz-target/in/fail01_EXCLUDE.json
+++ b/simd-fuzz-target/in/fail01_EXCLUDE.json
@@ -1,1 +1,0 @@
-"A JSON payload should be an object or array, not a string."