--- a/src/avx2/stage1.rs
+++ b/src/avx2/stage1.rs
@@ -28,6 +28,8 @@ macro_rules! high_nibble_mask {
 
 pub const SIMDJSON_PADDING: usize = mem::size_of::<__m256i>();
 pub const SIMDINPUT_LENGTH: usize = 64;
+pub const SIMD_ZERO_COUNT: usize  = 32;
+pub const SIMD_ZEROS: [u8; SIMD_ZERO_COUNT] = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
 
 #[derive(Debug)]
 pub(crate) struct SimdInput {