--- a/src/lib.rs
+++ b/src/lib.rs
@@ -389,23 +389,23 @@ impl<'de> Deserializer<'de> {
 
         let len = input.len();
 
-        let buf_start: usize = input.as_ptr() as *const () as usize;
-        let needs_relocation = (buf_start + input.len()) % page_size::get() < SIMDJSON_PADDING;
+        // let buf_start: usize = input.as_ptr() as *const () as usize;
+        // let needs_relocation = (buf_start + input.len()) % page_size::get() < SIMDJSON_PADDING;
 
-        let s1_result: std::result::Result<Vec<u32>, ErrorType> = if needs_relocation {
-            let mut data: Vec<u8> = Vec::with_capacity(len + SIMDJSON_PADDING);
-            unsafe {
-                data.set_len(len + 1);
-                data.as_mut_slice()
-                    .get_unchecked_mut(0..len)
-                    .clone_from_slice(input);
-                *(data.get_unchecked_mut(len)) = 0;
-                data.set_len(len);
-                Deserializer::find_structural_bits(&data)
-            }
-        } else {
-            unsafe { Deserializer::find_structural_bits(input) }
+        let mut buffer: Vec<u8> = Vec::with_capacity(len + SIMDJSON_PADDING * 2);
+        unsafe {
+            buffer.set_len(len + 1);
+            buffer
+                .as_mut_slice()
+                .get_unchecked_mut(0..len)
+                .clone_from_slice(input);
+            *(buffer.get_unchecked_mut(len)) = 0;
+            buffer.set_len(len);
         };
+
+        let s1_result: std::result::Result<Vec<u32>, ErrorType> =
+            unsafe { Deserializer::find_structural_bits(&buffer) };
+
         let structural_indexes = match s1_result {
             Ok(i) => i,
             Err(t) => {
@@ -413,7 +413,7 @@ impl<'de> Deserializer<'de> {
             }
         };
 
-        let tape = Deserializer::build_tape(input, &structural_indexes)?;
+        let tape = Deserializer::build_tape(input, buffer, &structural_indexes)?;
 
         Ok(Deserializer { tape, idx: 0 })
     }