--- a/src/lib.rs
+++ b/src/lib.rs
@@ -378,11 +378,29 @@ impl<'de> Deserializer<'de> {
     fn raw_error(idx: usize, c: char, error: ErrorType) -> Error {
         Error::new(idx, c, error)
     }
+
     // By convention, `Deserializer` constructors are named like `from_xyz`.
     // That way basic use cases are satisfied by something like
     // `serde_json::from_str(...)` while advanced use cases that require a
     // deserializer can make one with `serde_json::Deserializer::from_str(...)`.
     pub fn from_slice(input: &'de mut [u8]) -> Result<Self> {
+        let len = input.len();
+
+        let mut string_buffer: Vec<u8> = Vec::with_capacity(len + SIMDJSON_PADDING);
+        unsafe {
+            string_buffer.set_len(len + SIMDJSON_PADDING);
+        };
+
+        Deserializer::from_slice_with_buffer(input, &mut string_buffer)
+    }
+
+    // By convention, `Deserializer` constructors are named like `from_xyz`.
+    // That way basic use cases are satisfied by something like
+    // `serde_json::from_str(...)` while advanced use cases that require a
+    // deserializer can make one with `serde_json::Deserializer::from_str(...)`.
+    // this takes an additional buffer to be (re) used for temporary string copying
+
+    pub fn from_slice_with_buffer(input: &'de mut [u8], string_buffer: &mut [u8]) -> Result<Self> {
         // We have to pick an initial size of the structural indexes.
         // 6 is a heuristic that seems to work well for the benchmark
         // data and limit re-allocation frequency.
@@ -391,11 +409,13 @@ impl<'de> Deserializer<'de> {
 
         // let buf_start: usize = input.as_ptr() as *const () as usize;
         // let needs_relocation = (buf_start + input.len()) % page_size::get() < SIMDJSON_PADDING;
-
         let mut buffer: Vec<u8> = Vec::with_capacity(len + SIMDJSON_PADDING * 2);
-        let align = buffer.as_slice().as_ptr().align_offset(SIMDJSON_PADDING);
+
+        let align = buffer
+            .as_slice()
+            .as_ptr()
+            .align_offset(SIMDJSON_PADDING / 2);
         unsafe {
-            buffer.set_len(len + align + 1);
             buffer
                 .as_mut_slice()
                 .get_unchecked_mut(align..align + len)
@@ -414,7 +434,8 @@ impl<'de> Deserializer<'de> {
             }
         };
 
-        let tape = Deserializer::build_tape(input, &buffer[align..], &structural_indexes)?;
+        let tape =
+            Deserializer::build_tape(input, &buffer[align..], string_buffer, &structural_indexes)?;
 
         Ok(Deserializer { tape, idx: 0 })
     }