--- a/src/numberparse.rs
+++ b/src/numberparse.rs
@@ -112,9 +112,9 @@ fn is_made_of_eight_digits_fast(chars: &[u8]) -> bool {
     // return (( val & 0xF0F0F0F0F0F0F0F0 ) == 0x3030303030303030)
     //  && (( (val + 0x0606060606060606) & 0xF0F0F0F0F0F0F0F0 ) ==
     //  0x3030303030303030);
-    (((val & 0xF0F0_F0F0_F0F0_F0F0)
+    ((val & 0xF0F0_F0F0_F0F0_F0F0)
         | (((val.wrapping_add(0x0606_0606_0606_0606)) & 0xF0F0_F0F0_F0F0_F0F0) >> 4))
-        == 0x3333_3333_3333_3333)
+        == 0x3333_3333_3333_3333
 }
 
 #[cfg_attr(not(feature = "no-inline"), inline)]