--- a/tests/jsonchecker.rs
+++ b/tests/jsonchecker.rs
@@ -202,6 +202,6 @@ crash!(crash000033);
 crash!(crash000034);
 crash!(crash000035);
 crash!(crash000036);
-// crash!(crash000037);
+crash!(crash000037);
 // crash!(crash000038);
 // crash!(crash000039);