… .github/workflows/quality.yaml
… codecov.yml
… data/crash/crash000037.json
… fuzz/Cargo.toml
… simd-fuzz-target/Cargo.toml
… simd-fuzz-target/Makefile
… simd-fuzz-target/in/fail02.json
… simd-fuzz-target/in/fail03.json
… simd-fuzz-target/in/fail04.json
… simd-fuzz-target/in/fail05.json
… simd-fuzz-target/in/fail06.json
… simd-fuzz-target/in/fail07.json
… simd-fuzz-target/in/fail08.json
… simd-fuzz-target/in/fail09.json
… simd-fuzz-target/in/fail10.json
… simd-fuzz-target/in/fail11.json
… simd-fuzz-target/in/fail12.json
… simd-fuzz-target/in/fail13.json
… simd-fuzz-target/in/fail14.json
… simd-fuzz-target/in/fail15.json
… simd-fuzz-target/in/fail16.json
… simd-fuzz-target/in/fail17.json
… simd-fuzz-target/in/fail18_EXCLUDE.json
… simd-fuzz-target/in/fail19.json
… simd-fuzz-target/in/fail20.json
… simd-fuzz-target/in/fail21.json
… simd-fuzz-target/in/fail22.json
… simd-fuzz-target/in/fail23.json
… simd-fuzz-target/in/fail24.json
… simd-fuzz-target/in/fail25.json