Update to new fuzzer code for #95