@Licenser 2020-01-28
Initial perf check on changes to fix issues found with the new fuzzer:

note so great

## master
```
======= simd_json ======== parse|stringify ===== parse|stringify ====
data/canada.json         530 MB/s   420 MB/s   710 MB/s   330 MB/s
data/citm_catalog.json  1120 MB/s   480 MB/s  1540 MB/s   600 MB/s
data/twitter.json        920 MB/s   600 MB/s   990 MB/s   650 MB/s
data/log.json           1080 MB/s  1080 MB/s  1080 MB/s  1080 MB/s

===== simd_json_tape ===== parse|stringify ===== parse|stringify ====
data/canada.json         770 MB/s                    
data/citm_catalog.json  1900 MB/s                    
data/twitter.json       1670 MB/s                    
data/log.json           2170 MB/s                    
```

## this
```
======= simd_json ======== parse|stringify ===== parse|stringify ====
data/canada.json         500 MB/s   430 MB/s   670 MB/s   330 MB/s
data/citm_catalog.json  1050 MB/s   490 MB/s  1460 MB/s   620 MB/s
data/twitter.json        910 MB/s   620 MB/s   990 MB/s   650 MB/s
data/log.json           1080 MB/s  1080 MB/s  1080 MB/s  1080 MB/s

===== simd_json_tape ===== parse|stringify ===== parse|stringify ====
data/canada.json         740 MB/s                    
data/citm_catalog.json  1790 MB/s                    
data/twitter.json       1650 MB/s                    
data/log.json           2170 MB/s                    
```