@Licenser 2020-01-28
<pre>                                    Instructions             Cache.              Cycle/byte      
        Name           Cycles    Normal.     Branch     Misses   References    Best       Avg    
apache_builds            537705    1356915     172392        797       4305      4.190      4.225
apache_builds(+/-)   <font color="#4E9A06">   -9.975%</font> <font color="#4E9A06">   -9.243%</font> <font color="#4E9A06">   -5.063%</font> <font color="#CC0000">   70.891%</font> <font color="#CC0000">   70.407%</font> <font color="#4E9A06">   -3.516%</font> <font color="#4E9A06">   -9.975%</font>
canada                 18268717   53939041    7628342      18128      50422      8.073      8.116
canada(+/-)          <font color="#C4A000">    0.983%</font> <font color="#4E9A06">   -1.380%</font> <font color="#4E9A06">   -1.078%</font> <font color="#CC0000">   44.583%</font> <font color="#CC0000">   58.185%</font> <font color="#CC0000">    1.432%</font> <font color="#C4A000">    0.983%</font>
citm_catalog            7099823   18328666    2447906       8413      33624      4.073      4.111
citm_catalog(+/-)    <font color="#4E9A06">   -2.014%</font> <font color="#4E9A06">   -7.403%</font> <font color="#4E9A06">   -7.837%</font> <font color="#CC0000">   77.368%</font> <font color="#CC0000">   21.782%</font> <font color="#4E9A06">   -1.265%</font> <font color="#4E9A06">   -2.014%</font>
github_events            236860     543599      62042        400       1788      3.587      3.637
github_events(+/-)   <font color="#4E9A06">   -5.775%</font> <font color="#4E9A06">   -7.210%</font> <font color="#4E9A06">   -5.820%</font> <font color="#CC0000">   67.000%</font> <font color="#CC0000">   43.792%</font> <font color="#4E9A06">   -1.490%</font> <font color="#4E9A06">   -5.775%</font>
gsoc-2018               8031694   15905348    1443474      32090      87390      2.382      2.413
gsoc-2018(+/-)       <font color="#CC0000">    6.438%</font> <font color="#4E9A06">   -5.487%</font> <font color="#4E9A06">  -16.243%</font> <font color="#CC0000">   87.865%</font> <font color="#CC0000">   83.404%</font> <font color="#CC0000">   10.142%</font> <font color="#CC0000">    6.438%</font>
instruments             1352269    2982151     394612       1371      10838      6.080      6.137
instruments(+/-)     <font color="#CC0000">    3.222%</font> <font color="#4E9A06">   -2.253%</font> <font color="#C4A000">    0.342%</font> <font color="#CC0000">   78.848%</font> <font color="#CC0000">   62.991%</font> <font color="#CC0000">    4.650%</font> <font color="#CC0000">    3.222%</font>
log                       12919      24338       2893         11         71      5.608      5.940
log(+/-)             <font color="#C4A000">    0.612%</font> <font color="#4E9A06">   -6.028%</font> <font color="#4E9A06">   -3.249%</font> <font color="#CC0000">   81.818%</font> <font color="#CC0000">   47.887%</font> <font color="#C4A000">    0.590%</font> <font color="#C4A000">    0.612%</font>
marine_ik              29166886   71431965   10320335      51062     134342      9.718      9.776
marine_ik(+/-)       <font color="#CC0000">    3.925%</font> <font color="#4E9A06">   -1.378%</font> <font color="#4E9A06">   -1.309%</font> <font color="#CC0000">   62.191%</font> <font color="#CC0000">   62.030%</font> <font color="#CC0000">    3.769%</font> <font color="#CC0000">    3.925%</font>
mesh                    6443221   16584212    2362871       4278      22146      8.863      8.904
mesh(+/-)            <font color="#CC0000">    1.170%</font> <font color="#C4A000">    0.662%</font> <font color="#4E9A06">   -1.180%</font> <font color="#CC0000">   76.858%</font> <font color="#CC0000">   75.156%</font> <font color="#CC0000">    1.773%</font> <font color="#CC0000">    1.170%</font>
mesh.pretty             8430988   22055619    2935841       7571      31026      5.308      5.345
mesh.pretty(+/-)     <font color="#CC0000">    5.592%</font> <font color="#CC0000">    1.710%</font> <font color="#C4A000">   -0.840%</font> <font color="#CC0000">   74.772%</font> <font color="#CC0000">   76.500%</font> <font color="#CC0000">    5.520%</font> <font color="#CC0000">    5.592%</font>
numbers                  965364    2496051     311521        912       5568      6.398      6.430
numbers(+/-)         <font color="#4E9A06">   -1.476%</font> <font color="#C4A000">    0.931%</font> <font color="#4E9A06">   -2.803%</font> <font color="#CC0000">   73.794%</font> <font color="#CC0000">   75.952%</font> <font color="#CC0000">    2.181%</font> <font color="#4E9A06">   -1.476%</font>
random                  3402648    8355062    1052801       3066      19965      6.629      6.666
random(+/-)          <font color="#4E9A06">   -6.729%</font> <font color="#4E9A06">  -10.234%</font> <font color="#4E9A06">   -4.260%</font> <font color="#CC0000">   86.367%</font> <font color="#CC0000">   37.145%</font> <font color="#4E9A06">   -5.993%</font> <font color="#4E9A06">   -6.729%</font>
twitter                 3212697    6728927     767723       3790      29549      5.047      5.087
twitter(+/-)         <font color="#CC0000">    3.554%</font> <font color="#4E9A06">   -4.728%</font> <font color="#4E9A06">   -2.327%</font> <font color="#CC0000">   84.697%</font> <font color="#CC0000">   67.830%</font> <font color="#CC0000">    4.624%</font> <font color="#CC0000">    3.554%</font>
twitterescaped          3910963    8354387    1015571       3307      27423      6.915      6.954
twitterescaped(+/-)  <font color="#C4A000">   -0.652%</font> <font color="#4E9A06">   -8.045%</font> <font color="#4E9A06">   -9.677%</font> <font color="#CC0000">   85.455%</font> <font color="#CC0000">   64.668%</font> <font color="#C4A000">    0.002%</font> <font color="#C4A000">   -0.652%</font>
update-center           2762077    6104194     773926       3221      22675      5.144      5.180
update-center(+/-)   <font color="#4E9A06">   -1.921%</font> <font color="#4E9A06">  -10.788%</font> <font color="#4E9A06">   -5.789%</font> <font color="#CC0000">   85.408%</font> <font color="#CC0000">   68.419%</font> <font color="#C4A000">   -1.000%</font> <font color="#4E9A06">   -1.921%</font>
</pre>